requirements.txt
markdown>=3.4.0
Jinja2>=3.1.0
python-frontmatter>=1.0.0
PyYAML>=6.0
